Make more explicit the contents of `./.vscode/JuliaSysimage.toml`
I am missing a lot of detail in the explanation for how to add a script to be run in the automated sysimage creation. This PR contains my best guess as to the answers I am looking for in order to demonstrate what I mean. The intention is for someone who knows the answers to correct me ^_^